test(di): include botocore in exploration testing (#13853)
We include botocore in the boto3 exploration testing suite to increase
the amount of code that gets instrumented. We also introduce
configuration options to limit the amount of discovered code that gets
randomly instrumented. This is a solution that has been adopted to allow
tests suite that are heavily affected by the instrumentation to complete
within a reasonable amount of time. The instrumentation of `botocore`
through the `boto3` run falls into this category, so we set the
instrumentation rate to 5%.

Refs: https://datadoghq.atlassian.net/browse/DEBUG-4105

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [ ] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)