feat: poc implementation of the OTel Metrics API (#13780)
## Motivation
We want to implement support for the OTel Metrics API across our
libraries. There is an
[RFC](https://docs.google.com/document/d/1Zkpk6QdlRQ67459fqHk8_SlJKb3-ZqRg-3juLHtWBtU/edit?usp=sharing)
in development as well as
[system-tests](https://github.com/DataDog/system-tests/pull/4790) to
define that behavior, and this POC aims to deliver the necessary code to
implement that.

For API usage, see the `utils/build/docker/python/flask/app.py` file in
the referenced system-tests PR.

Tracking Issue: [APMAPI-1322]

## Summary
This PR provides an opt-in implementation of the OTel Metrics API by
taking an optional dependency on the OpenTelemetry SDK packaage, so it
will only be enabled when both the flag is enabled and the dependency is
present. By bringing in the SDK component, we minimize API
incompatibilities for the time being and we cover the OTel Metrics API
that application / library developers expect:
- Get/Set a global `MeterProvider` instance (which provides access to
`Meter`s)
- Get a `Meter` with a given name+version
- Create synchronous/asynchronous instruments with a given name, kind,
unit, and description. This also allows users to directly record
measurements on the synchronous instruments.

All of this functionality is controlled by the environment variable
`DD_TRACE_OTEL_METRICS_ENABLED`, which is disabled by default.

## Testing
This PR adds unit tests for the configuration, ensuring that we do not
load any metrics SDK types unless the user has brought their SDK +
exporter AND `DD_TRACE_OTEL_METRICS_ENABLED=true`. In terms of
end-to-end testing, I am currently relying on the linked system-tests
PR, which this branch is currently passing.

## Improvements
This is still early in the development, so we can expect to implement
additional changes in the future, such as making the integration with
Datadog environment variables complete and setting metric preferences
that align with Datadog's processing. For now, I'd like to merge as-is.

## Related Changes
There are a number of changes that are adopted from
https://github.com/DataDog/dd-trace-py/pull/13831, including riotfile
changes and configuration test code. When one PR is merged, the other
should rebase to make merging easier.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)
- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
- - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
- - - - - - - - - - - -

[APMAPI-1322]:
https://datadoghq.atlassian.net/browse/APMAPI-1322?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ

---------

Co-authored-by: Munir Abdinur <munir.abdinur@datadoghq.com>
Co-authored-by: Munir Abdinur <munir_abdinur@hotmail.com>