chore(profiling): increase error tolerance for test_accuracy_stack_v2
A 7% error tolerance seems to not quite be enough, and this is flaky.
I don't think our accuracy is any better or worse than it's been in the
past. The test just has some inherent variance. Bump the threshold to
10% for now just to keep CI a little more green.

Longer term there might be ways we can improve this test. At first
glance it's surprising to see this much error: spend_cpu_2 spins for 2
seconds of CPU time, and so we assert that there are roughly 2 seconds
of CPU time attributed to that funciton in the profile. Yet we regularly
see much less then that. This might be in part because we're measuring
_process_ CPU time when we spin, yet there are other threads using CPU
time besides the one running spend_cpu_2, such as the stack sampler.
Maybe there's a way we can use _thread_ CPU time?