fix(di): dynamic function discovery fallback
We implement a dynamic function discovery fallback when the
function-from-code resolution via the GC fails. This can happen if the
target application has interacted with the GC, e.g. by freezing it at a
time that will prevent the current discovery from being able to resolve
the function from the referenced code object.