chore(profiling): de-flake test_accuracy_stack_v2 (#13944)
test_accuracy_stack_v2 is flaky. The test asserts that we attribute a
certain amount of CPU time to specific functions. These functions try to
use up a given amount of CPU time, using `time.process_time_ns`, which
counts the amount of CPU time used by the process, to track how much CPU
they've used. This is problematic because there are multiple threads in
the process: the thread running the test functions, the stack sampling
thread, and whatever other threads (telemetry?) `ddtrace` creates. We
should instead use `time.thread_time_ns`, which counts just the CPU time
used by the calling thread, to determine how much we spin in the tests.

This PR makes that change, and defensively increases the error tolerance
to 10%. At least testing locally, I didn't see any failures after many
runs with a tolerance as low as 5%. But it's annoying if this test
flakes. We can keep the error tolerance a little higher for now in case
there's yet another reason for significant variance I haven't
anticipated.