ci: add global lock detection
We add a CI check to detect the presence of global locks in all the
modules loaded by the library. Relying on modules with internal global
lock introduces the risk of potential races with target applications
that might rely on the functionalities exposed by these modules. This
could result in a deadlock in case of forks.