fix(profiling): support for greenlets
We improve the error handling around greenlet tracking to catch any
unexpected errors and prevent the target application from failing to
spawn greenlets. One observed case of failure was due to the string
table failing to resolve a string (the greenlet name in this case) into
a cache key. We expect this kind of errors to be rare, but we add logic
to handle them should they occur.