chore(profiling): skip broken test_gunicorn on 3.8
This test is broken right now and we don't have a clear diagnosis. The
failure seems to consistenly be on 3.8 and the test appears mostly
stable on other versions. For now, let's skip it on 3.8 so that we don't
need to keep retrying the test while we diagnose the issue.