fix(profiling): make explicitly marked main packages "my code" (#13649)
If a user runs their code as a "main module" which they installed like a
library (`python -m my_program args...`), and expicilty specifies that
it's the main package via `DD_MAIN_PACKAGE`, we should consider code
from
that module "my code". Right now we don't, though, because we mark that
code as a "library" in the profiling code provenance info.  Check
whether the user specified a main package and make sure it gets marked
as "my code" by leaving the kind blank.