fix(tracer): compute trace times correctly in the presence of freezegun (#14036)
In the past, `freezegun` would cause span durations and telemetry data
to be wrongly computed due to the patching of functions imported from
the `time` module. PR #11406 introduced a `freezegun` integration that
modified `freezegun`'s list of ignored modules to include `ddtrace`. The
problem with this approach is that `freezegun` looks up the whole call
stack to check if it's being called from an ignored module, which means
that anything that is wrapped by a `ddtrace` function ends up being
ignored as well, including our own `unittest` wrappers.

This PR introduces a `Time` class that saves references to the builtin
time functions that won't be touched by `freezegun`, and changes the
methods used by spans and telemetry to use time functions from that
class. In this way, those methods are protected from the effects of
`freezegun` even if the integration is not enabled.

In a subsequent PR (#14037), the integration will be removed, thus
solving the problem with `unittest`.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)