fix(ci_visibility): do not propagate ddtrace logs to the root logger (#14121)
When pytest is run with log capture on (which is the default, unless
`-s` or similar is used), it replaces `sys.stdout` and `sys.stderr` with
temporary file descriptors that are closed at the end of the test
session. If a custom logger is created referencing one of those streams,
any logs emitted after the end of the test session will try to write to
the closed file descriptor, leading to an ` I/O operation on closed
file` error. See #12018 for example cases.

Annoyingly, this happens even if ddtrace is not explicitly enabled in
pytest with `--ddtrace`, because before we even have the opportunity to
parse command line arguments, `ddtrace/__init__.py` will load, which
sets up a bunch of things, including loggers.

At the beginning of a pytest session, we call a
`take_over_logger_stream_handler()` function which sets up the `ddtrace`
logger specifically for test runs. This PR:
- Changes `take_over_logger_stream_handler()` to not propagate logs from
the `ddtrace` handler to the root handler. This means the `ddtrace`
handler will not be affected by custom loggers by default.

- Makes `take_over_logger_stream_handler()` run even if `--ddtrace` is
not enabled. This is not the ideal solution in the sky: ideally we would
not enable ddtrace features at all when `--ddtrace` is not activated,
however this is a bigger change to the way ddtrace works at the moment,
so this PR focuses on avoiding the logging issue.

Finally, this change makes the current tests that use `caplog` to check
for logs emitted by `ddtrace` to break, because `caplog` depends on the
logs being propagated to the root logger to capture them, so this PR
adds a wrapping `caplog` fixture in our `conftest.py` that re-enables
log propagation for those tests. (The alternative would be changing
every `caplog.at_level(lvl)` call in our tests to `caplog.at_level(lvl,
"ddtrace")`, but there are _a lot_ of them.)

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)