refactor: native forksafe threads
We refactor the native periodict thread implementation to be
fork-safe, the sense that all such threads that are running at the
time of a fork are automatically stopped before the fork, then
restarted after it. We also take care to avoid stopping and
restarting threads in the parent process if we detect an immediate
call to fork again.