chore(profiling): drop windows support (#14140)
It has come to my attention that supporting windows takes quite an
effort

- #12073: Started on Jan 31, but finally merged on April 1. To iterate
on this PR, I had to spin up a windows vm, and iterate there. Still, it
was not as smooth as my devex on linux workspace or on corp laptop.
- #13984: This PR changes the way dd-trace-py depends on libdatadog, and
even without windows support for profiling it's already more than 50
commits and taking 2 engineers' time. And this PR is not the first
attempt, previous attempts are #13978, #13603.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)