fix(ci_visibility): handle `skipif` with non-positional arguments (#14150)
The ddtrace pytest plugin currently assumes that the `skipif` marker
will be provided the condition as a positional argument. It breaks if no
condition is provided, or if the condition is passed as a keyword
argument.

This leads to the `_pytest_runtest_protocol_pre_yield()` call raising an
exception, which leaves `coverage_collector` unset, which leads to the
call to `_pytest_runtest_protocol_post_yield()` to fail. If the test was
not finished during `pytest_runtest_protocol` (for example, because the
`pytest-rerunfailures` or `flaky` plugins are in use, which causes
ddtrace to skip its own `pytest_runtest_protocol`), the test span will
be left unfinished and with a `fail` test status.

This PR adds handling for keyword and absent `skipif` conditions, and
also makes sure `coverage_collector` is set even if
`_pytest_runtest_protocol_pre_yield()` fails.

- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

- [x] Reviewer has checked that all the criteria below are met
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)

(cherry picked from commit 13fcc382d05c271855addcbbae3db9ba4477100f)