chore: stricter wrapping context wrap check
We make the check on whehter a function is wrapped with a universal
wrapping context stricter. This will now check that we have actual
bytecode wrapping, as the attribute check alone is not strict enough.
There are examples of frameworks that might copy the state of wrapped
functions over to closures, which would lead the current implementation
to believe that a function is wrapped when in reality it is not.