feat(llmobs): allow span processor to return None to omit spans (#14260)
Allow users to omit any traces they want by returning none in the
user_span_processor method.

Motivation:
I am trying to prevent some auto traces from cluttering my Datadog
observability dashboards. These traces create noise and make it harder
to focus on the more critical traces. My LLM observability overview is
filled with what is clustered as empty input which is incorrect. e.g.
all embedding traces are are just a spammer.

To do that, I am using the new span_processor, and the new way to omit
specific spans, will be to return null by the relevant span processor.

I have added a test for that.

The only risk is that there is no indication that span was omitted but I
was afraid that a debug log would be too spammy. But please let me know
if that will help. Also there are some telemetric collected about sent
span which might also need to be changed, will be great to hear your
thoughts

Co-authored-by: amirbenami <amir@linxsecurity.io>