chore(aap): appsec span processor without import cycles (#14244)
This PR introduces:
- remove of the `_appsec_processor` in the tracer. The appsec processor
is now part of the list of regular processors. This fixes several
problems.
- change the logic order of how appsec is enabled and loaded

This is removing 2 static import cycles, only 2 left (none with appsec).

Context: Before that PR, the appsec product was loaded through the
tracer, by creating an instance of AppSecSpanProcessor. After that PR,
the appsec product or the remote config layer will register the
AppSecSpanProcessor into the list of regular span processors of the
tracer.


**Bonus**:
As the appsec processor is not reset anymore at fork time by the respawn
of the tracer (because it does not depend on the tracer object anymore),
it looks like it's also fixing the gap time we could have where appsec
was back to factory setting at work respawn time.

APPSEC-57505

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)