ci(langchain): try and fix flaky tests (#14282)
Another best-effort attempt to remove langchain test flakiness. This is
done by remove all in-process VCR usage for request recording, as well
as openai mocks using `respx`, instead relying entirely on the
testagent's VCR capabilities to record cassettes to the mounted
directory, so no http patching is done in-process.

To go with this, I've removed some tests that are no longer compatible
without updating the testagent. These are mainly
1. ai21 tests
2. cohere tests
3. bedrock tests
4. pinecone tests

of which each only had 1 or 2. to remedy this, either in this PR or
another, we can:
1. replace them with nicer `langchain-genai` tests to verify other
langchain providers
2. update the testagent to include these endpoints. i didn't do it for
this PR as i just wanted to see if removing in-process mocking worked.
should be easy enough for ai21 and cohere, but bedrock will be tricky
because of request signing and verification when generating the
cassettes. for pinecone, i attempted to use the in-memory vectorstore
for simplicity, although i guess we don't actually patch it.

## Checklist
- [x] PR author has checked that all the criteria below are met
- The PR description includes an overview of the change
- The PR description articulates the motivation for the change
- The change includes tests OR the PR description describes a testing
strategy
- The PR description notes risks associated with the change, if any
- Newly-added code is easy to change
- The change follows the [library release note
guidelines](https://ddtrace.readthedocs.io/en/stable/releasenotes.html)
- The change includes or references documentation updates if necessary
- Backport labels are set (if
[applicable](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting))

## Reviewer Checklist
- [x] Reviewer has checked that all the criteria below are met 
- Title is accurate
- All changes are related to the pull request's stated goal
- Avoids breaking
[API](https://ddtrace.readthedocs.io/en/stable/versioning.html#interfaces)
changes
- Testing strategy adequately addresses listed risks
- Newly-added code is easy to change
- Release note makes sense to a user of the library
- If necessary, author has acknowledged and discussed the performance
implications of this PR as reported in the benchmarks PR comment
- Backport labels are set in a manner that is consistent with the
[release branch maintenance
policy](https://ddtrace.readthedocs.io/en/latest/contributing.html#backporting)